backend/pyproject.toml
[project]
name = "euclid-backend"
version = "0.1.0"
description = "Euclid Pretty Pics Portal Backend"
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "httpx",
    "python-dotenv"
]
